+++ tests/tests_app/structures/test_structures.py
@@ -46,6 +46,7 @@ def test_dict():
             "_restarting": False,
             "_display_name": "",
             "_internal_ip": "",
+            "_public_ip": "",
             "_cloud_compute": {
                 "type": "__cloud_compute__",
                 "name": "cpu-small",
@@ -80,6 +81,7 @@ def test_dict():
             "_restarting": False,
             "_display_name": "",
             "_internal_ip": "",
+            "_public_ip": "",
             "_cloud_compute": {
                 "type": "__cloud_compute__",
                 "name": "cpu-small",
@@ -114,6 +116,7 @@ def test_dict():
             "_restarting": False,
             "_display_name": "",
             "_internal_ip": "",
+            "_public_ip": "",
             "_cloud_compute": {
                 "type": "__cloud_compute__",
                 "name": "cpu-small",
@@ -199,6 +202,7 @@ def test_list():
             "_paths": {},
             "_restarting": False,
             "_internal_ip": "",
+            "_public_ip": "",
             "_display_name": "",
             "_cloud_compute": {
                 "type": "__cloud_compute__",
@@ -233,6 +237,7 @@ def test_list():
             "_paths": {},
             "_restarting": False,
             "_internal_ip": "",
+            "_public_ip": "",
             "_display_name": "",
             "_cloud_compute": {
                 "type": "__cloud_compute__",
@@ -262,6 +267,7 @@ def test_list():
             "_paths": {},
             "_restarting": False,
             "_internal_ip": "",
+            "_public_ip": "",
             "_display_name": "",
             "_cloud_compute": {
                 "type": "__cloud_compute__",